requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
icalendar==7.0.3
idna==3.10
isodate==0.7.2
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
propcache==0.3.2
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
TatSu==5.13.1
types-python-dateutil==2.9.0.20250822
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
